+++ src/roller/table.ts
@@ -1,12 +1,12 @@
-import { MarkdownRenderer, Notice, Pos } from "obsidian";
+import { MarkdownRenderer, Notice, Pos,TFile } from "obsidian";
 
 import { TABLE_REGEX } from "src/utils/constants";
-import { StackRoller } from ".";
+import { LinkRoller, StackRoller, TagRoller } from ".";
 import { GenericFileRoller } from "./roller";
 
 class SubRollerResult {
-    result: string="";
-    combinedTooltip: string="";
+    result: string = "";
+    combinedTooltip: string = "";
 }
 
 export class TableRoller extends GenericFileRoller<string> {
@@ -57,23 +57,33 @@ export class TableRoller extends GenericFileRoller<string> {
         if (this.plugin.data.displayResultsInline) {
             result.unshift(this.inlineText);
         }
-        MarkdownRenderer.renderMarkdown(
+        const div = createSpan();
+        await MarkdownRenderer.renderMarkdown(
             result.join(""),
-            this.resultEl.createSpan("embedded-table-result"),
+            div,
             this.source,
             null
         );
+        const resultEl = this.resultEl.createSpan("embedded-table-result");
+        if (
+            div.childElementCount == 1 &&
+            div.firstElementChild instanceof HTMLParagraphElement
+        ) {
+            resultEl.append(...Array.from(div.firstElementChild.childNodes));
+        } else {
+            resultEl.append(...Array.from(div.childNodes));
+        }
     }
 
     prettify(input: string): string {
-        const specialChars = /(.*?)(\(|\)|;|\|\|)(.*)/
+        const specialChars = /(.*?)(\(|\)|;|\|\|)(.*)/;
         const tab = "\t";
         let tabCount = 0;
-        let output:string = "";
+        let output: string = "";
 
-        let remaining:string = input;
+        let remaining: string = input;
         let matches: RegExpMatchArray;
-        while (matches = remaining.match(specialChars)) {
+        while ((matches = remaining.match(specialChars))) {
             let [, beforeSpecial, special, afterSpecial] = matches;
             output += beforeSpecial;
             if (special == ")") {
@@ -81,19 +91,16 @@ export class TableRoller extends GenericFileRoller<string> {
                 output += "\n";
                 output += tab.repeat(tabCount);
                 output += ")";
-            }
-            else {
+            } else {
                 if (special == "(") {
                     tabCount++;
                     output += "(";
-                }
-                else if (special == ";") {
+                } else if (special == ";") {
                     output += ",";
-                }
-                else if (special == "||") {
+                } else if (special == "||") {
                     output += "|";
                 }
-                output += "\n"
+                output += "\n";
                 output += tab.repeat(tabCount);
             }
             remaining = afterSpecial;
@@ -107,8 +114,7 @@ export class TableRoller extends GenericFileRoller<string> {
         let res: SubRollerResult = new SubRollerResult();
         if (typeof input === "number") {
             res.result = input.toString();
-        }
-        else {
+        } else {
             res.result = input;
         }
         let subTooltips: string[] = [];
@@ -124,20 +130,29 @@ export class TableRoller extends GenericFileRoller<string> {
                 const formula = foundRoller[1].trim();
 
                 // Create sub roller with formula
-                const subRoller = await this.plugin.getRoller(formula, this.source);
+                const subRoller = await this.plugin.getRoller(
+                    formula,
+                    this.source
+                );
                 // Roll it
                 await subRoller.roll();
                 // Get sub result
                 const rollerResult = await this.getSubResult(subRoller.result);
 
+                let result: string;
+                if ((rollerResult.result as any) instanceof TFile) {
+                    result = (rollerResult.result as unknown as TFile).basename;
+                } else {
+                    result = rollerResult.result;
+                }
+
                 // Replace dice block by sub result
-                res.result = res.result.replace(foundRoller[0], rollerResult.result);
+                res.result = res.result.replace(foundRoller[0], result);
 
                 // Update tooltip
                 if (subRoller instanceof TableRoller) {
                     subTooltips.push(subRoller.combinedTooltip);
-                }
-                else {
+                } else {
                     const [top, bottom] = subRoller.tooltip.split("\n");
                     subTooltips.push(top + " --> " + bottom);
                 }
@@ -177,9 +192,9 @@ export class TableRoller extends GenericFileRoller<string> {
         }
 
         for (let i = 0; i < this.rolls; i++) {
-            let subTooltip:string = "";
+            let subTooltip: string = "";
             let subResult: SubRollerResult;
-            let selectedOption:string = "";
+            let selectedOption: string = "";
 
             if (this.isLookup) {
                 const result = await this.lookupRoller.roll();
@@ -189,14 +204,27 @@ export class TableRoller extends GenericFileRoller<string> {
                         (result >= range[0] && range[1] >= result)
                 );
                 if (option) {
-                    subTooltip = this.lookupRoller.original.trim() + " --> " + `${this.lookupRoller.resultText}${this.header ? " | " + this.header : ""}`.trim();
+                    subTooltip =
+                        this.lookupRoller.original.trim() +
+                        " --> " +
+                        `${this.lookupRoller.resultText}${
+                            this.header ? " | " + this.header : ""
+                        }`.trim();
                     selectedOption = option[1];
                 }
-            }
-            else {
+            } else {
                 const options = [...this.options];
-                const randomRowNumber = this.getRandomBetween(0, options.length - 1);
-                subTooltip = options.length + " rows" + " --> " + "[row " + (randomRowNumber+1) + "]";
+                const randomRowNumber = this.getRandomBetween(
+                    0,
+                    options.length - 1
+                );
+                subTooltip =
+                    options.length +
+                    " rows" +
+                    " --> " +
+                    "[row " +
+                    (randomRowNumber + 1) +
+                    "]";
                 selectedOption = options[randomRowNumber];
             }